[Removed] part of the cookies logic to use serverside storage and localStorage instead
