+++ Client/package.json
@@ -22,7 +22,6 @@
     "@fortawesome/free-regular-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/vue-fontawesome": "^3.0.5",
-    "@tinymce/tinymce-vue": "^5.1.1",
     "axios": "^1.6.3",
     "pinia": "^2.1.7",
     "sass": "^1.69.6",
